\name{betfair-package}
\alias{betfair-package}
\alias{betfair}
\docType{package}
\title{
Simple Betfair API for R.
}
\description{
}
\details{
\tabular{ll}{
Package: \tab betfair\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-16\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
colin@betwise.co.uk 
}
\references{
}
