% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_methods.R
\name{betfair}
\alias{betfair}
\title{betfair - Betfair Environment w/ methods}
\usage{
betfair(usr, pwd, key)
}
\arguments{
\item{usr}{Betfair username}

\item{pwd}{Betfair password}

\item{key}{Betfair API key see \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Application+Keys}{developer.betfair.com}}
}
\value{
returns environment with functions for the various methods available for
Betfair's API
}
\description{
Supply valid credentials, username, password and API key (see \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Application+Keys}{developer.betfair.com}),
returning a environment with methods (see Methods section below) for returning data from
Betfair's API.
}
\details{
filters can be used in a number of available methods, these filters are
to be supplied as a list, the \link{marketFilter} function provides assistance in
building the list.
}
\section{Methods}{

\describe{
     \item{\code{competitions(filter = marketFilter())}}{ Retrieve data about the
     different competitions with current markets, see \link{competitions}.}
     \item{\code{countries(filter = marketFilter())}}{ Retrieve data about the different countries
     hosting events, see \link{countries}.}
     \item{\code{events(filter = marketFilter())}}{ Retrieve data about the different events, see
     \link{events}.}
     \item{\code{eventTypes(filter = marketFilter())}}{ Retrieve data about the different event types,
     ie. sports, see \link{eventTypes}.}
     \item{\code{venues(filter = marketFilter())}}{ Retrieve data about the venues hosting racing,
     see \link{venues}.}
     \item{\code{login(usr, pwd, key)}}{ Login in, a session token will be
     returned, over-writing the previous token when \code{betfair(usr, pwd, key)}
     was used.}
     \item{\code{session()}}{ Print details about the session, including login
     in details and session token.}
     \item{\code{marketBook(marketIds = list(), priceProjection = NULL, orderProjection = NULL,
     matchProjection = NULL)}}{ Retrieve dynamic data about markets. Data includes
     prices, the status of the market, the status of the selections, the traded
     volume, and the status of any orders you have in the market.}
     \item{\code{marketCatalogue(filter = list(), marketProjection = "EVENT", sort = NULL,
     maxResults = 50, keepRules = FALSE)}}{ Retrieve data about the different types
     of markets, see \link{marketCatalogue}.}
     \item{\code{marketTypes(filter = marketFilter())}}{ Retrieve data about the different types of
     markets, see \link{marketTypes}.}
     \item{\code{cancelOrders(..., marketId = NA)}}{ Cancel existing orders, leave function blank
     to cancel all existing orders, or use combination or marketId and ... to target specific orders,
     see \link{cancel_inst} to target specific order}
     \item{\code{placeOrders(marketId, selectionId, orderType = "LIMIT",
     handicap = NULL, side = "BACK", limitOrder = bf_helpers$limitOrder(), limitOnCloseOrder = bf_helpers$limitOnCloseOrder(),
     marketOnCloseOrder = NULL)}}{ Place a bet, requires a marketId (see \link{marketCatalogue})
     and selectionId (see \link{marketBook}),}
     \item{\code{replaceOrders(..., marketId)}}{ Replace existing orders, requires marketIds and
     unique betIds with new prices, use \link{replace_inst} to target specific bets}
}
}
\examples{
\dontrun{
bf <- betfair(usr = "username", pwd = "password", key = "API_key")

# to view available methods simply print the bf
bf

# to use a method, for example competitions
bf$competitions(filter = list())

# to use a method with a filter to retrieve just football
bf$competitions(filter = marketFilter(eventTypeIds = 1))
}
}

