% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_login.R
\name{betfair_login}
\alias{betfair_login}
\title{login to Betfair}
\usage{
betfair_login(usr, pwd, key)
}
\arguments{
\item{usr}{Betfair username}

\item{pwd}{Betfair password}

\item{key}{Betfair API Key, see \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Application+Keys}{developer.betfair.com}}
}
\value{
list with class \code{betfaiR_login} which contains the parameters
entered into the function, the response (includes a session token) and the
required headers for querying the API
}
\description{
login to Betfair to receive a session token
}
\details{
function is invoked when calling \link{betfair}
}

