\name{betfair_odds}
\alias{betfair_odds}
\docType{data}
\title{betfair_odds
}
\description{
Betfair betfair_odds
}
\usage{data(betfair_odds)}
\format{
  The format is:
 num [1:350] 1.01 1.02 1.03 1.04 1.05 1.06 1.07 1.08 1.09 1.1 ...
}
\details{
}
\source{
}
\references{
}
\examples{
data(betfair_odds)
## maybe str(betfair_odds) ; plot(betfair_odds) ...
}
\keyword{datasets}
