% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_parse.R
\name{betfair_parse}
\alias{betfair_parse}
\title{betfair_parse}
\usage{
betfair_parse(res, ...)
}
\arguments{
\item{res}{response to be parsed into dataframes}
}
\description{
convert response from the various \link{betfair} methods, tries
to convert response into useful format, ie. dataframes with zero nested elements
}
\details{
due to the various methods returning different responses varying in
detail and complexity (testing all combinations is tricky), so this function
may not be that reliable at the moment
}

