% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_post.R
\name{betfair_POST}
\alias{betfair_POST}
\title{betfair_POST}
\usage{
betfair_POST(body, headers)
}
\arguments{
\item{body}{request body, json object containing data about the method and
params}

\item{headers}{headers, list object containing session token and api key}
}
\description{
POST request to Betfair's Exchange API, requires users to prepare
a request using \link{base_request}, \link{betfair_request} and to have login
credentials from \link{betfair_login}.
}

