% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_request.R
\name{betfair_request}
\alias{betfair_request}
\title{create Betfair request}
\usage{
betfair_request(x, ...)
}
\description{
convert and format request list into valid JSON object for Betfair's
API. Requires users to build a list first using \link{base_request}
}
\details{
to use this function, the object should have one of the following classes,
\strong{competitions}, \strong{countries}, \strong{events}, \strong{eventTypes},
\strong{marketCatalogue}, \strong{marketTypes} and \strong{venues}.
}

