% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_theme.R
\name{betfair_theme}
\alias{betfair_theme}
\title{betfair_theme}
\usage{
betfair_theme(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\description{
ggplot2 theme for betfaiR package plots, theme is used to plot the
output of the \link{statement} method
}

