\name{betStatus}
\alias{betStatus}
\docType{data}
\title{betStatus
}
\description{
Betfair betStatus
}
\usage{data(betStatus)}
\format{
  The format is:
List of 5
 $ : chr "Cancelled"
 $ : chr "Lapsed"
 $ : chr "Matched"
 $ : chr "Settled"
 $ : chr "Unmatched"
}
\details{
}
\source{
}
\references{
}
\examples{
data(betStatus)
## maybe str(betStatus) ; plot(betStatus) ...
}
\keyword{datasets}
