% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_account.R
\name{bf_account}
\alias{bf_account}
\title{bf_account - Betfair Account Environment w/ methods}
\usage{
bf_account(usr, pwd, key)
}
\arguments{
\item{usr}{Betfair username}

\item{pwd}{Betfair password}

\item{key}{Betfair API key see \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Application+Keys}{developer.betfair.com}}
}
\value{
returns environment with functions for various methods available for Betfair's
Account API
}
\description{
Supply valid credentials, username, password and API key (see \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Application+Keys}{developer.betfair.com}),
returning an environment with methods (see Methods section below) for returning data
about your account
}
\section{Methods}{

\describe{
     \item{\code{funds()}}{ Retrieve available funds, see \link{funds}}
     \item{\code{details()}}{ Retreive account details, see \link{details}}
     \item{\code{transfer(from = "uk", to = "aus", amount = 2)}}{ Transfer funds from one wallet
     to another, see \link{transfer}}
}
}

