% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_check.R
\name{bf_check}
\alias{bf_check}
\title{check response is valid}
\usage{
bf_check(res, method = "competitions")
}
\arguments{
\item{res}{response object}

\item{method}{the method the response object came from, see methods section on
\link{betfair} page, or \link{bf_basic_req}.}
}
\description{
check for any errors in the response from \link{bf_post}
}
\details{
errors may be invalid JSON, missing parameters, or if the response contains
zero results, otherwise return the results with a class that corresponds to
the method (see \link{betfair}) that can then be parsed into a dataframe by
\link{bf_parse}.
}

