% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\docType{data}
\name{bf_helpers}
\alias{bf_helpers}
\title{betfair helper functions}
\format{\preformatted{<environment: 0x000000000a7e9630> 
}}
\usage{
bf_helpers
}
\description{
a environment with collection of helper functions that are used
in conjunction with the various methods available.
}
\section{Helpers}{

\describe{
     \item{\code{prepare(marketId, selectionId, orderType = "LIMIT",
     handicap = "0", side = "BACK",limitOrder = bf_helpers$limitOrder(),
     limitOnCloseOrder = bf_helpers$limitOnCloseOrder(),
     marketOnCloseOrder = list())}}{ function used inside \link{prepareOrders}
     method to help construct an order, uses other helper functions \code{limitOrder}
     or \code{limitOnCloseOrder}, see \link{prepare}}
     \item{\code{limitOrder(size = 2, price = NULL, persistenceType = "LAPSE")}}{
     helper function used in conjunction with \code{prepare} when an orderType is
     \strong{LIMIT}, see \link{limitOrder}}
     \item{\code{limitOnCloseOrder(size = 2, price = NULL)}}{ helper function
     used in conjunction with \code{prepare}, when an orderType is \strong{LIMIT_ON_CLOSE},
     see \link{limitOnCloseOrder}}
     \item{\code{cancel(..., marketId = NA)}}{ helper function used by the
     the \link{cancelOrders} method to cancel all bets if no params are entered,
     or by using \code{cancel_inst} to target individual bets, or target markets
     via the marketId param, see \link{cancel}}
     \item{\code{cancel_inst(betId = NA, size = NA)}}{ helper function used with
     \code{cancel} to cancel individual bets, can be used multiple times inside
     \code{cancel}, see \link{cancel_inst}}
     \item{\code{replace_inst(betId, newPrice)}}{ helper function used by the
     \link{replaceOrders} method, betId and newPrice are required, see \link{replace_inst}}
}
}
\keyword{datasets}

