% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_post.R
\name{bf_post}
\alias{bf_post}
\title{post a request to Betfair}
\usage{
bf_post(body, headers, dom = "uk")
}
\arguments{
\item{body}{request body, json object containing data about the method and
params}

\item{headers}{headers, list object containing session token and api key}

\item{dom}{Betfair domain, default is "uk" for the European exchange, change to
"aus" for Australian exchange}
}
\description{
POST request to Betfair's Exchange API, requires users to prepare
a request using \link{bf_basic_req}, \link{bf_request} and to have login
credentials from \link{bf_login}.
}

