% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{cancel}
\alias{cancel}
\title{cancel}
\arguments{
\item{...}{multiple cancel orders instructions, use \link{cancelInstruction}}

\item{marketId}{market id}
}
\value{
list
}
\description{
cancel orders function
}

