% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{cancel_inst}
\alias{cancel_inst}
\title{cancel bet instructions}
\usage{
cancel_inst(betId = NA, size = NA)
}
\arguments{
\item{betId}{unique betId}

\item{size}{reduction size}
}
\value{
list with cancel instructions
}
\description{
cancel individual orders using betIds
}

