% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cancel_orders}
\alias{cancel_orders}
\title{cancel_orders}
\usage{
cancel_orders(marketId = NULL, ...)
}
\arguments{
\item{marketId}{market id}

\item{...}{multiple cancel orders instructions, use \link{cancelInstruction}}
}
\value{
list
}
\description{
cancel orders function
}

