% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cancelInstruction}
\alias{cancelInstruction}
\title{cancelInstruction}
\usage{
cancelInstruction(betId = NULL, size = NULL)
}
\arguments{
\item{betId}{unique betId}

\item{size}{reduction size}
}
\description{
cancel individual orders using betIds
}

