% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{cancelOrders}
\alias{cancelOrders}
\title{cancelOrders method}
\arguments{
\item{...}{target specific bets to cancel, use \link{cancel_inst} to enter
one or more bets, \code{cancel_inst} requires \strong{betId}, and \strong{size}
can be used to cancel part of an order}

\item{marketId}{target bets in a specific market}
}
\value{
list with details about the cancel bets, whether the cancellation was successful
}
\description{
\code{cancelOrders} function cancels all or specific bets waiting to
be matched
}

