% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{collect_data}
\alias{collect_data}
\title{collect data}
\usage{
collect_data(bf, marketIds, marketIds_file = "marketIds",
  priceProjection = c("EX_ALL_OFFERS", "EX_TRADED"))
}
\arguments{
\item{bf}{betfair environment, returned from the \link{betfair} function}

\item{marketIds}{ids of markets to retrieve data for}

\item{marketIds_file}{filename for markets collecting data from}

\item{priceProjection}{what data to return via the \link{marketBook} function}
}
\value{
the function saves data returned from the Betfair API into RDS files
named after the market, for example market_1.121212121.RDS.
}
\description{
helper function to be used within an R script that is part of a
scheduled task.
}
\details{
This function will take a \link{betfair} environment, a
collection of marketIds and priceProjection parameters and retrieve data at
intervals set up by the task, and keep appending new data returned by the
\link{marketBook} method to any previous data returned.  The function will also
add a new variable called \code{collectedAt} to each list, which means cross
market comparisons are possible.  It is a good idea to also store/save data
returned by the \link{marketCatalogue} method which will contain data about the
runners in the markets you are collecting data from.
}

