% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{countries}
\alias{countries}
\title{countries method}
\arguments{
\item{filter}{list to select desired markets, see \link{marketFilter},
or visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-MarketFilter}{developer.betfair.com},
for the different options.}
}
\value{
a dataframe of countries with markets selected by the \code{filter}
parameter, with the following variables:
\itemize{
     \item{\code{countryCode} the ISO-2 code for the event, ISO-2 codes are
     available on \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{wiki}}
     \item{\code{marketCount} number of markets associated with this competition}
}
}
\description{
countries associated with markets
}

