% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{currentOrders}
\alias{currentOrders}
\title{currentOrders method}
\arguments{
\item{betId}{unique bet Id}

\item{marketId}{unique market Id}

\item{orderProjection}{default of \strong{ALL} returns all unsettled positions,
matched or partially matched, change to \strong{EXECUTABLE} to filter for orders
with a portion remaining, or \strong{EXECUTION_COMPLETE} for orders which have
been filled, see orderProjection section in \link{bettingEnums} for more details}

\item{from}{date filter, string in yyyy-mm-dd format}

\item{to}{date filter, string in yyyy-mm-dd format}

\item{orderBy}{how to order orders, default orders by when they were placed, see
orderBy section in \link{bettingEnums} for more details}

\item{sort}{how to sort results, see sortDir section in \link{bettingEnums} for
more details}

\item{fromRecord}{specifies the first record to be returned, records start at
index zero (not one)}

\item{count}{specifies how many records are returned from the index position set
by \code{fromRecord}, there is a limit of 1000.}
}
\value{
list with data about individual orders
}
\description{
\code{currentOrders} function allows users to retrieve data about
any unsettled bets they have open.  Use the parameters to filter to specific
markets, or leave the function empty to return \strong{all} open positions
}

