% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_account.R
\name{details}
\alias{details}
\title{details}
\value{
dataframe with the following variables, \code{currencyCode}, \code{firstName},
\code{lastName}, \code{localeCode}, \code{region}, \code{timezone}, \code{discountRate},
\code{pointsBalance}, \code{countryCode}
}
\description{
get data about the account you have signed into
}
