% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{events}
\alias{events}
\title{events method}
\arguments{
\item{filter}{list to select desired markets, see \link{marketFilter},
or visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-MarketFilter}{developer.betfair.com},
for the different options.}
}
\value{
a dataframe of events (ie. Manchester United vs Arsenal) associated
with markets selected by the \code{filter} parameter, with the following variables:
\itemize{
     \item{\code{event_id} event id}
     \item{\code{event_name} event name}
     \item{\code{event_countryCode} ISO-2 country code for the event}
     \item{\code{event_timezone} the timezone in which the event is taking place}
     \item{\code{event_venue} the event venue (if applicable)}
     \item{\code{event_openDate} the scheduled start Date of rhe event}
     \item{\code{marketCount} number of markets associated with this competition}
}
}
\description{
events associated with markets
}
\examples{
\dontrun{

# login
bf <- betfair(usr = "username", pwd = "password", key = "API_key")

# return all events, probably best to store as dataframe, likely exceed 1000 rows
events_df <- bf$events()

# return football events
# using eventType id learned by using bf$eventTypes()
football <- bf$events(filter = marketFilter(eventTypeIds = 1))
}
}

