% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{eventTypes}
\alias{eventTypes}
\title{eventTypes method}
\arguments{
\item{filter}{list to select desired markets, see \link{marketFilter},
or visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-MarketFilter}{developer.betfair.com},
for the different options.}
}
\value{
a dataframe of event types associated with markets selected by the
\code{filter} parameter, with the following variables:
\itemize{
     \item{\code{eventType_id} eventType id}
     \item{\code{eventType_name} eventType name}
     \item{\code{marketCount} number of markets associated with this competition}
}
}
\description{
event types associated with markets
}

