\name{Exchange}
\alias{Exchange}
\docType{data}
\title{Exchange
}
\description{
Betfair Exchange
}
\usage{data(Exchange)}
\format{
  The format is:
 chr "https://api.betfair.com/exchange/v5/BFExchangeService"
}
\details{
}
\source{
}
\references{
}
\examples{
data(Exchange)
## maybe str(Exchange) ; plot(Exchange) ...
}
\keyword{datasets}
