\name{ExchangeAU}
\alias{ExchangeAU}
\docType{data}
\title{ExchangeAU
}
\description{
Betfair ExchangeAU
}
\usage{data(ExchangeAU)}
\format{
  The format is:
 chr "https://api-au.betfair.com/exchange/v5/BFExchangeService"
}
\details{
}
\source{
}
\references{
}
\examples{
data(ExchangeAU)
## maybe str(ExchangeAU) ; plot(ExchangeAU) ...
}
\keyword{datasets}
