% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_account.R
\name{funds}
\alias{funds}
\title{funds}
\value{
dataframe with the following variables, \code{availableToBetBalance},
\code{exposure}, \code{retainedCommission}, \code{exposureLimit}, \code{discountRate},
\code{pointsBalance} and \code{wallet}
}
\description{
get available funds in your account
}
