% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{limitOnCloseOrder}
\alias{limitOnCloseOrder}
\title{limitOnCloseOrder}
\usage{
limitOnCloseOrder(size = 2, price = NULL)
}
\arguments{
\item{size}{the size of the bet. \strong{Note:} for a market type of EACH_WAY,
the total stake is 2*size.}

\item{price}{the limit price}
}
\value{
list with bet order
}
\description{
helper function to prepare a bet, used in conjunction with \link{prepare_order}
}

