% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{limitOrder}
\alias{limitOrder}
\title{limitOrder}
\usage{
limitOrder(size = 2, price = NULL, persistenceType = "LAPSE")
}
\arguments{
\item{size}{the size of the bet. \strong{Note:} for a market type of EACH_WAY,
the total stake is 2*size.}

\item{price}{the limit price}

\item{persistenceType}{what to do with the order at turn in-play, three choices,
     \strong{LAPSE} - lapse the order when the market goes in play,
     \strong{PERSIST} - persist the order in-play, the bet will be placed
     automatically into the in-play market at the start of the event,
     \strong{MARKET_ON_CLOSE} - put the order into the auction (SP) at turn
     in-play.}
}
\value{
list with bet order
}
\description{
helper function to prepare a bet, used in conjunction with \link{prepare_order}
}

