% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{marketBook}
\alias{marketBook}
\title{marketBook method}
\arguments{
\item{marketIds}{(string) one or more market Ids. The number of markets returned
depends on the amount of data you request via the price projection.}

\item{priceProjection}{the projection of price data you want to receive in the
response, see section
\strong{PriceProjection} in \link{bettingEnums}.}

\item{orderProjection}{the orders you want to receive in the response, see
section \strong{OrderProjection} in \link{BettingEnums}.}

\item{matchProjection}{if you ask for orders, specifies the representation of
matches, see section \strong{MatchProjection} in \link{BettingEnums}.}
}
\description{
dynamic data about markets. Data includes prices, the status of the
market, the status of selections, the traded volume, and the status of any orders
you have placed in the market.
}
\details{
Seperate requests should be made for ACTIVE and CLOSED markets.
Requests that include both ACTIVE and CLOSED markets will only return those
markets that are active.

Data request limits apply to requests using \code{marketBook} that include price
or order projections. Calls to \code{marketBook} should be made up to a maximum of
5 times per second to a single marketId.

\strong{BEST PRACTICE}: those seeking to use \code{marketBook} to obtain price,
volume, unmatched orders and matched position in a single operation should
provide an \code{orderProjection} of \strong{EXECUTABLE} in their \code{marketBook}
request and receive all unmatched orders and the aggregated matched volume from
all orders.  The level of matched volume aggregation (\code{matchProjection})
should be \strong{ROLLED_UP_BY_AVG_PRICE} or \strong{ROLLED_UP_BY_PRICE}, the
former being preferred. This provides a single call in which you can track prices,
traded volume, unmatched orders, and your evolving matched position with a reasonably
fixed, minimally sized response.
}

