% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marketFilter.R
\name{marketFilter}
\alias{marketFilter}
\title{helper function to construct list to filter data for various available methods}
\usage{
marketFilter(bspOnly = NULL, competitionIds = NULL, eventIds = NULL,
  eventTypeIds = NULL, exchangeIds = NULL, inPlayOnly = NULL,
  marketBettingTypes = NULL, marketCountries = NULL, marketIds = NULL,
  from = NULL, to = NULL, marketTypeCodes = NULL, textQuery = NULL,
  turnInPlayEnabled = NULL, venues = NULL, withOrders = NULL)
}
\arguments{
\item{bspOnly}{\code{(boolean)} Restrict to BSP (Betfair Starting Price)
markets only, TRUE, or non-BSP markets, FALSE. If not specified then returns
both BSP and non-BSP markets.}

\item{competitionIds}{\code{(string)} Restrict markets by the competitions
associated with the market.}

\item{eventIds}{\code{(string)} Restrict markets by the event Id associated
with the market.}

\item{eventTypeIds}{\code{(numeric)} Restrict markets by event type associated
with the market, ie. Football, Hockey, etc.}

\item{exchangeIds}{\code{(numeric)} Restrict markets by the Exchange where the
market operates. As of Jun 11, 2015, \strong{not available}.}

\item{inPlayOnly}{\code{boolean} Restrict to markets that are currently in
play, TRUE, or not currently in play, FALSE. If not specified returns both.}

\item{marketBettingTypes}{see section \strong{MarketBettingType} in \link{BettingEnums}
or \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Enums#MarketBettingType}{developer.betfair.com}
Restrict to markets that match the betting type of the market, ie. Odds,
Asian Handicap Singles, Asian Handicap Doubles.}

\item{marketCountries}{\code{string} Restrict to markets that are in the
specified country or countries.}

\item{marketIds}{\code{numeric} Restrict markets by the market Id associated
with the market.}

\item{from}{\code{string} start time after specified date, format date as "YYYY-MM-DD"
or with a time "YYYY-MM-DD 00:00:00".}

\item{to}{\code{string} start time before specified date, format date as "YYYY-MM-DD"
or with a time "YYYY-MM-DD 00:00:00".}

\item{marketTypeCodes}{\code{numeric} Restrict to markets that match the type
of the market, ie. MATCH_ODDS, HALF_TIME_SCORE. You should use this instead
of relying on the market name as the market type codes are the same in all
locales.}

\item{textQuery}{\code{string} Restrict markets by any text associated with
the market such as Name, Event, Competition, etc. You can include a wildcard
(*) character, as long as it is not the first character.}

\item{turnInPlayEnabled}{\code{boolean} Restrict to markets that will turn in
play, TRUE, or will not turn in play, FALSE. If not specified returns both.}

\item{venues}{\code{string} Restrict markets by the venue associated with the
market. Currently, as of Jun 11, 2015, only Horse Racing markets have venues.}

\item{withOrders}{see section \strong{OrderStatus} in \link{BettingEnums}
or \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Enums#OrderStatus}{developer.betfair.com}
Restrict to markets that I have one or more orders in these status.}
}
\description{
build a market filter object given the parameters available, this
object can be used in the following methods \link{competitions}, \link{countries},
\link{events}, \link{eventTypes}, \link{marketCatalogue}, and \link{marketTypes}
}

