% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{marketPnL}
\alias{marketPnL}
\title{marketPnL method}
\arguments{
\item{marketId}{target bets in a specific market}

\item{settled}{include settled bets}

\item{bsp}{include bets that will be made at BSP}

\item{NET}{include commission}
}
\value{
list with data about the market and any profit and loss for selections
}
\description{
\code{marketPnL} function returns data about the specified market
and the current profit and/or loss for each of the selections in that market
}

