% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{marketTypes}
\alias{marketTypes}
\title{marketTypes method}
\arguments{
\item{filter}{list to select desired markets, see \link{marketFilter},
or visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-MarketFilter}{developer.betfair.com},
for the different options.}
}
\value{
a dataframe of market types (ie. MATCH_ODDS, NEXT_GOAL) associated
with markets selected by the \code{filter} parameter, with the following variables:
\itemize{
     \item{\code{marketType} the market type}
     \item{\code{marketCount} number of markets associated with this competition}
}
}
\description{
marketTypes associated with markets
}

