% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{placeOrders}
\alias{placeOrders}
\title{placeOrders method}
\arguments{
\item{marketId}{market id into which the bet will be placed}

\item{selectionId}{selection id, the unique id for selection in the market, this
will be a horse, a team, etc, these ids can be found using the \link{marketCatalogue}
and \link{marketBook} functions}

\item{orderType}{the type of order, \strong{LIMIT} a normal exchange order for
immediate execution, \strong{LIMIT_ON_CLOSE} limit order for the auction (SP bet)
\strong{MARKET_ON_CLOSE} market order for the auction (SP)}

\item{handicap}{The handicap associated with the runner in case of Asian handicap markets, null otherwise.}

\item{side}{Back or Lay order}

\item{order}{the type of order, this can be three types and inputs depend on
parameter \strong{orderType}.  If orderType is \strong{LIMIT} then use the
\link{limitOrder} function to construct order.  If orderType is \strong{LIMIT_ON_CLOSE}
use the \link{limitOnCloseOrder} function to construct the order.  If orderType is
\strong{MARKET_ON_CLOSE} then a simple list consisting of one element called
\strong{liability} should be supplied}
}
\value{
list with details about the bet, the market the bet was made in, the prices
attained, whether it was successful or failed
}
\description{
placeOrders function allows bets to be placed.
}

