% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{prepare}
\alias{prepare}
\title{prepare bet for Betfair}
\usage{
prepare(marketId, selectionId, orderType = "LIMIT", handicap = "0",
  side = "BACK", order = NULL)
}
\arguments{
\item{marketId}{the unique market id, starts with \code{1.} for UK exchange,
and \code{2.} for Aus exchange}

\item{selectionId}{the selection (runner) unique id}

\item{orderType}{type of order, \strong{LIMIT} (default) for immediate execution,
\strong{LIMIT_ON_CLOSE} limit order for the auction (SP), \strong{MARKET_ON_CLOSE}
market order for the auction (SP)}

\item{handicap}{the handicap associated with the runner (\code{selectionId}) in
case of Asian Handicap markets}

\item{side}{"BACK" (default) or "LAY"}

\item{order}{the type of order, this can be three types and inputs depend on
parameter \strong{orderType}.  If orderType is \strong{LIMIT} then use the
\link{limitOrder} function to construct order.  If orderType is \strong{LIMIT_ON_CLOSE}
use the \link{limitOnCloseOrder} function to construct the order.  If orderType is
\strong{MARKET_ON_CLOSE} then a simple list consisting of one element called
\strong{liability} should be supplied}
}
\value{
returns a list with an order for Betfair
}
\description{
prepare a bet for \link{placeOrders} method
}

