% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_order}
\alias{prepare_order}
\title{prepare bet for Betfair}
\usage{
prepare_order(marketId, selectionId, orderType = "LIMIT", handicap = "0",
  side = "BACK", limitOrder = limitOrder(),
  limitOnCloseOrder = limitOnCloseOrder(), marketOnCloseOrder = list())
}
\arguments{
\item{marketId}{the unique market id, starts with \code{1.} for UK exchange,
and \code{2.} for Aus exchange}

\item{selectionId}{the selection (runner) unique id}

\item{orderType}{type of order, \strong{LIMIT} (default) for immediate execution,
\strong{LIMIT_ON_CLOSE} limit order for the auction (SP), \strong{MARKET_ON_CLOSE}
market order for the auction (SP)}

\item{handicap}{the handicap associated with the runner (\code{selectionId}) in
case of Asian Handicap markets}

\item{side}{"BACK" (default) or "LAY"}

\item{limitOrder}{a simple exchange bet for immediate execution, a list consisting of
\itemize{
     \item \strong{size} the size of the bet. \strong{Note:} for a market type
     of EACH_WAY, the total stake is 2*size.
     \item \strong{price} the limit price
     \item \strong{persistenceType} what to do with the order at turn in-play,
     three choices, \strong{LAPSE} - lapse the order when the market goes in play,
     \strong{PERSIST} - persist the order in-play, the bet will be placed automatically
     into the in-play market at the start of the event, \strong{MARKET_ON_CLOSE} -
     put the order into the auction (SP) at turn in-play.
}}

\item{limitOnCloseOrder}{bets are matched if, and only if, the returned starting
price is better than a specified price, a list consisting of
\itemize{
     \item \strong{liability} the size of the bet
     \item \strong{price} the limit price of the bet
}}

\item{marketOnCloseOrder}{Bets remain unmatched until the market is reconciled,
hey are matched and settled at a price that is representative of the market at
the point the market is turned in-play, a list consisting of
\itemize{
     \item \strong{liability} the size of the bet
}}
}
\value{
returns a list with an order for Betfair
}
\description{
prepare a bet for \link{placeOrders} method
}

