% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{replace_inst}
\alias{replace_inst}
\title{replace bet instructions}
\usage{
replace_inst(betId, newPrice)
}
\arguments{
\item{betId}{unique betId (required)}

\item{newPrice}{new price to strike bet (required)}
}
\value{
list with replace instructions
}
\description{
replace individual orders using betIds
}

