% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{replaceInstructions}
\alias{replaceInstructions}
\title{replaceInstructions}
\usage{
replaceInstructions(betId, newPrice)
}
\arguments{
\item{betId}{unique betId (required)}

\item{newPrice}{new price to strike bet (required)}
}
\description{
replace individual orders using betIds
}

