% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{replaceOrders}
\alias{replaceOrders}
\title{replaceOrders method}
\arguments{
\item{...}{target specific bets to cancel, use \link{replace_inst} to enter
one or more bets, \code{bf_helpers$replace_inst} requires \strong{betId}, and
\strong{newPrice}}

\item{marketId}{target bets in a specific market}
}
\value{
list with details about the replaced bets, whether operation was successful
}
\description{
\code{replaceOrders} function replaces specific bets that are waiting to
be matched
}

