% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_account.R
\name{statement}
\alias{statement}
\title{statement}
\arguments{
\item{after}{filter records after this date, date should be in YYYY-MM-DD format}

\item{before}{filter records before this date, date should be in YYYY-MM-DD format}

\item{n}{number of records}

\item{wallet}{filter according to wallet, default returns both, enter "UK" or
"AUSTRALIAN"}

\item{include}{filter according to transaction type}
}
\value{
list of transactions, one element per transaction, this list can be
passed to \code{print}, \code{summary} or \code{plot}
}
\description{
return account statement with all or filtered transactions,
including deposits, withdrawals, bets, transfers, etc
}

