% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betfair_account.R
\name{transfer}
\alias{transfer}
\title{transfer}
\arguments{
\item{from}{wallet, "UK" (default) or "AUSTRALIAN"}

\item{to}{wallet, "UK" (default) or "AUSTRALIAN"}

\item{amount}{amount to transfer, minimum of 2 (default)}
}
\value{
transaction id
}
\description{
transfer funds between UK and Australian wallets
}

