% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{update_inst}
\alias{update_inst}
\title{update bet instructions}
\arguments{
\item{betId}{unique betId (required)}

\item{persistenceType}{new instructions for what to do with order at in play}
}
\description{
update individual orders with new instructions for what to do when
the market goes in play
}

