% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{venues}
\alias{venues}
\title{venues method}
\arguments{
\item{filter}{list to select desired markets, see \link{marketFilter},
or visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-MarketFilter}{developer.betfair.com},
for the different options.}
}
\value{
a dataframe of venues (ie. Cheltenham, Ascot) associated with markets
selected by the \code{filter} parameter, only Horse Racing markets are
associated with a venue.
}
\description{
venues associated with markets
}

