% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{cancel}
\alias{cancel}
\title{helper function for constructing request for canceling orders}
\usage{
cancel(..., marketId = NA)
}
\arguments{
\item{...}{multiple cancel orders instructions, use \link{cancel_inst}}

\item{marketId}{market id}
}
\value{
list
}
\description{
cancel orders function
}

