% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{cancel_inst}
\alias{cancel_inst}
\title{cancel bet instructions}
\arguments{
\item{betId}{unique betId}

\item{size}{reduction size}
}
\description{
cancel individual orders using betIds
}

