% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betfair_helpers.R
\name{cleared}
\alias{cleared}
\title{cleared}
\arguments{
\item{betStatus}{filter based on how the bet was settled, one of \strong{SETTLED},
\strong{VOIDED}, \strong{LAPSED}, or \strong{CANCELLED}}

\item{eventTypeIds}{retrieve bets based on sport}

\item{eventIds}{retrieve bets based on event}

\item{marketIds}{retrieve bets based on markets}

\item{runnerIds}{retrieve bets on specific runners}

\item{betIds}{unique bets}

\item{side}{\strong{BACK} or \strong{LAY} orders}

\item{from}{filter according to date, format should be yyyy-mm-dd}

\item{to}{filter according to date, format should be yyyy-mm-dd}
}
\value{
list with parameters to filter cleared bets
}
\description{
helper function used by the \link{clearedOrders} method to build
request object for finding settled orders
}

