% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{marketCatalogue}
\alias{marketCatalogue}
\title{marketCatalogue method}
\arguments{
\item{filter}{list to select desired markets, see \strong{marketFilter} section below
or visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-MarketFilter}{developer.betfair.com},
for the different options.}

\item{marketProjection}{the type and amount of data returned about the market, default of \strong{EVENT}, see
section \strong{marketProjection} in \link{BettingEnums}, or visit
\href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Enums#BettingEnums-MarketProjection}{developer.betfair.com}}

\item{sort}{the order of results, see section \strong{sort} in \link{BettingEnums}, or
visit \href{https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Enums#BettingEnums-MarketSort}{developer.betfair.com}}

\item{maxResults}{limit on the number of results returned, default of 1}

\item{keepRules}{when \strong{MARKET_DESCRIPTION} is entered into the \strong{marketProjection}
parameter, it returns a large amount of text containing the rules of the market, by default these
are discarded, change \code{keepRules = TRUE} to keep.}
}
\value{
a dataframe of markets that does not change (or changes very rarely), will
contain data about the market ids, selections and other info, with a selection
of the following variables:
}
\description{
marketCatalogue associated with markets, you use \code{marketCatalogue}
to retrieve the id of a market, the names of selections and other information about
the market
}
\details{
You use \code{marketCatalogue} to retrieve the name of the market, the
names of selections and other information about markets.  Market Data Request
Limits apply to requests.
}

