% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{placeOrders}
\alias{placeOrders}
\title{placeOrders method}
\arguments{
\item{marketId}{market id into which the bet will be placed}

\item{selectionId}{selection id, the unique id for selection in the market, this
will be a horse, a team, etc, these ids can be found using the \link{marketCatalogue}
and \link{marketBook} functions}

\item{orderType}{the type of order, \strong{LIMIT} a normal exchange order for
immediate execution, \strong{LIMIT_ON_CLOSE} limit order for the auction (SP bet)
\strong{MARKET_ON_CLOSE} market order for the auction (SP)}

\item{handicap}{}

\item{side}{\strong{BACK} or \strong{LAY}}

\item{limitOrder}{list with details of the bet, required if \strong{orderType} is
\strong{LIMIT}, use \link{limitOrder} function to help construct correct list}

\item{limitOnCloseOrder}{list with details of the bet, required if \strong{orderType}
is \strong{LIMIT_ON_CLOSE}, use \link{marketOnCloseOrder} function to help
construct correct list}

\item{marketOnCloseOrder}{single number, the size of your bet, which will be
struck at Betfair SP}
}
\value{
list with details about the bet, the market the bet was made in, the prices
attained, whether it was successful or failed
}
\description{
\code{placeOrders} function allows bets to be placed.
}

