% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{updateOrders}
\alias{updateOrders}
\title{updateOrders}
\arguments{
\item{...}{target specific bets to update, use \link{update_inst} to enter one
or more bets, \code{bf_helpers$update_inst} requires \strong{betId} and
\strong{persistenceType}}

\item{marketId}{target bets in a specific market}
}
\value{
list with details about the updated bets
}
\description{
\code{updateOrders} function updates instructions for specific bets
and how they are handled when the market goes in play
}

